+++ clients/admin-ui/src/features/system/SystemsTable.tsx
@@ -40,7 +40,10 @@ const SystemsTable = () => {
   return (
     <>
       {messageContext}
-      <Flex justify="space-between" className="mb-4">
+      <Flex
+        justify="space-between"
+        className="sticky -top-6 z-10 -mt-6 bg-white py-4"
+      >
         <DebouncedSearchInput
           value={searchQuery}
           onChange={updateSearch}