+++ clients/admin-ui/src/pages/data-discovery/action-center/website/[monitorId]/[systemId]/index.tsx
@@ -2,7 +2,7 @@ import { NextPage } from "next";
 import { useRouter } from "next/router";
 import { useEffect } from "react";
 
-import FixedLayout from "~/features/common/FixedLayout";
+import Layout from "~/features/common/Layout";
 import {
   ACTION_CENTER_ROUTE,
   UNCATEGORIZED_SEGMENT,
@@ -36,7 +36,7 @@ const MonitorResultAssets: NextPage = () => {
   }, [systemResults, router, monitorId]);
 
   return (
-    <FixedLayout title="Action center - Discovered assets">
+    <Layout title="Action center - Discovered assets">
       <PageHeader
         heading="Action center"
         breadcrumbItems={[
@@ -52,13 +52,14 @@ const MonitorResultAssets: NextPage = () => {
                 : system?.name,
           },
         ]}
+        isSticky={false}
       />
       <DiscoveredAssetsTable
         monitorId={monitorId}
         systemId={systemId}
         consentStatus={system?.consent_status}
       />
-    </FixedLayout>
+    </Layout>
   );
 };